removes litgpt dependency from thunder/benchmarks/utils.py
